** moved test suite up a level in repo.
