+++ tests/test_data/cnn_regression_test_data.py
