+++ tests/test_gpt.py
@@ -6,7 +6,7 @@ from nlpmodels.models import gpt
 from nlpmodels.utils import utils
 from nlpmodels.utils.gpt_batch import GPTBatch
 from nlpmodels.utils.vocabulary import NLPVocabulary
-from nlpmodels.tests.test_data import gpt_regression_test_data
+from tests.test_data import gpt_regression_test_data
 
 utils.set_seed_everywhere()
