+++ tests/test_vocabulary.py
